+++ crux/src/Crux/Model.hs
@@ -34,13 +34,24 @@ emptyModelView = ModelView $ MapF.empty
 toDouble :: Rational -> Double
 toDouble = fromRational
 
-
-showBVLiteral :: (1 <= w) => NatRepr w -> BV w -> String
-showBVLiteral w bv =
+showBVLiteralSigned :: (1 <= w) => NatRepr w -> BV w -> String
+showBVLiteralSigned w bv =
     (if x < 0 then "-0x" else "0x") ++ N.showHex i (if natValue w == 64 then "L" else "")
   where
   x = BV.asSigned w bv
   i = abs x
+
+showBVLiteralUnsigned :: (1 <= w) => NatRepr w -> BV w -> String
+showBVLiteralUnsigned w bv =
+    "0x" ++ N.showHex i (if natValue w == 64 then "L" else "")
+  where
+  i = BV.asUnsigned bv
+
+showBVLiteralDecimal :: (1 <= w) => NatRepr w -> BV w -> String
+showBVLiteralDecimal w bv =
+    show x
+  where
+  x = BV.asSigned w bv
 
 showFloatLiteral :: BigFloat -> String
 showFloatLiteral x
@@ -61,7 +72,7 @@ showDoubleLiteral x
 valsJS :: BaseTypeRepr ty -> Vals ty -> IO [JS]
 valsJS ty (Vals xs) =
   let showEnt = case ty of
-        BaseBVRepr n -> showEnt' (showBVLiteral n) n
+        BaseBVRepr n -> showBVEnt n
         BaseFloatRepr (FloatingPointPrecisionRepr eb sb)
           | Just Refl <- testEquality eb (knownNat @8)
           , Just Refl <- testEquality sb (knownNat @24)
@@ -85,6 +96,18 @@ valsJS ty (Vals xs) =
          , "val"  ~> jsStr (repr (entryValue e))
          , "bits" ~> jsStr (show n)
          ]
+
+  showBVEnt :: (1 <= w) => NatRepr w -> Entry (BV w) -> IO JS
+  showBVEnt n e = do
+    l <- fromMaybe jsNull <$> jsLoc (entryLoc e)
+    pure $ jsObj
+      [ "name"         ~> jsStr (entryName e)
+      , "loc"          ~> l
+      , "val"          ~> jsStr (showBVLiteralSigned n (entryValue e))
+      , "val-unsigned" ~> jsStr (showBVLiteralUnsigned n (entryValue e))
+      , "val-decimal"  ~> jsStr (showBVLiteralDecimal n (entryValue e))
+      , "bits"         ~> jsStr (show n)
+      ]
 
 modelJS :: ModelView -> IO JS
 modelJS m =